requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
passlib==1.7.4
prisma==0.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyhumps==3.8.0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.8.0
uvicorn==0.23.2
